src/requirements.txt
arrow==1.0.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
playsound==1.2.2
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
rq==1.7.0
rq-dashboard==0.6.1
schedule==1.0.0
six==1.15.0
tinydb==4.4.0
urllib3==1.26.3
Werkzeug==1.0.1
